Improved scripts and reorganized structure
Added additional functionality to both Unix and Windows scripts. Windows can now launch launch the web browser as part of the script. Also reorganized the file structure to clean up the root directory. Now only the necessities are in it, and the server and web app code are in the directory 'app' in order to reduce visual clutter.