Added the ability to track inactivity, improved README
Using last battle played, players can not be labelled as INACTIVE to assist in finding only active players to recruit. Currently they are displayed at the end. There is a plan to make them not displayable at all. Additionally, improved the README to outline what the user can do with the configuration file.